Optimize the functionality of downloading and saving files

+++ backend/web/js/pages/dropzone.init.js
@@ -89,16 +89,27 @@ function dropzoneInit(form, options) {
     });
 
     dropzone.on("success", function(file, response) {
-        if (typeof response == "object") {
+        // if (typeof response == "object") {
+        //     if (!response.data.success) {
+        //         if (!(response.data.errors === undefined || response.data.errors.length == 0)) {
+        //             const fileErrors = response.data.errors.files.find(el => el.filename === file.name)?.errors;
+        //             if (fileErrors) {
+        //                 addErrors(file, fileErrors);
+        //                 $(file.previewElement).find('[data-dz-uploadprogress]').removeClass('bg-success').addClass('bg-danger');
+        //             } else {
+        //                 this.removeFile(file);
+        //             }
+        //         }
+        //     } else {
+        //         this.removeFile(file);
+        //     }
+        // }
+        if (typeof response === "object" && response.data) {
             if (!response.data.success) {
-                if (!(response.data.errors === undefined || response.data.errors.length == 0)) {
-                    const fileErrors = response.data.errors.files.find(el => el.filename === file.name)?.errors;
-                    if (fileErrors) {
-                        addErrors(file, fileErrors);
-                        $(file.previewElement).find('[data-dz-uploadprogress]').removeClass('bg-success').addClass('bg-danger');
-                    } else {
-                        this.removeFile(file);
-                    }
+                const fileStatus = response.data.files[file.name];
+                if (fileStatus && fileStatus.status === 'error') {
+                    addErrors(file, fileStatus.errors);
+                    $(file.previewElement).find('[data-dz-uploadprogress]').removeClass('bg-success').addClass('bg-danger');
                 }
             } else {
                 this.removeFile(file);
@@ -107,21 +118,31 @@ function dropzoneInit(form, options) {
     });
     
     dropzone.on('error', function (file, response) {
-        if (typeof response == "object") {
-            if (response.data.errors) {
-                if (response.data.errors.message && typeof response.data.errors.message == 'string') {
-                    addError(file, response.data.errors.message);
-                } else {
-                    const fileErrors = response.data.errors.files.find(el => el.filename === file.name)?.errors;
-                    if (fileErrors) {
-                        addErrors(file, fileErrors);
-                    }
-                }
-            } 
-        } else {
-            addError(file, (typeof response == "string" ? response : 'Не удалось сохранить файл.'));
-        }
+        // if (typeof response == "object") {
+        //     if (response.data.errors) {
+        //         if (response.data.errors.message && typeof response.data.errors.message == 'string') {
+        //             addError(file, response.data.errors.message);
+        //         } else {
+        //             const fileErrors = response.data.errors.files.find(el => el.filename === file.name)?.errors;
+        //             if (fileErrors) {
+        //                 addErrors(file, fileErrors);
+        //             }
+        //         }
+        //     } 
+        // } else {
+        //     addError(file, (typeof response == "string" ? response : 'Не удалось сохранить файл.'));
+        // }
 
+        // $(file.previewElement).find('[data-dz-uploadprogress]').removeClass('bg-success').addClass('bg-danger');
+
+        if (typeof response === "object" && response.data) {
+            const fileStatus = response.data.files[file.name] || response.data.files['global'];
+            if (fileStatus && fileStatus.errors) {
+                addErrors(file, fileStatus.errors);
+            }
+        } else {
+            addError(file, (typeof response === "string" ? response : 'Не удалось сохранить файл.'));
+        }
         $(file.previewElement).find('[data-dz-uploadprogress]').removeClass('bg-success').addClass('bg-danger');
     });
